exl2/requirements.txt
torch>=1.13.0
transformers>=4.25.1
safetensors>=0.3.1
sentencepiece>=0.1.97
huggingface_hub>=0.16.4
exllamav2>=0.0.5
python-dotenv>=0.19.0